fix(harness): consolidate duplicated math scoring into shared module
_extract_boxed, _normalize_math, and _math_equiv were copy-pasted
across three files with no shared import. generation_benchmark's copy
was missing currency/unit normalization, \frac whitespace handling,
and mixed-number matching — causing benchmark scores to disagree on
identical model outputs.

Move the superset implementation (from bench_llm.py) into
harness/math_scoring.py. Import from the three consumers and delete
the local copies.
